+++ include/OptionalAPI.hpp
@@ -2,7 +2,7 @@
 
 #include "Branding.hpp"
 
-#include <Geode/Geode.hpp>
+#include <Geode/Result.hpp>
 
 #include <Geode/loader/Dispatch.hpp>
 
@@ -11,13 +11,11 @@
 #endif
 #define MY_MOD_ID "cheeseworks.moddevbranding"
 
-using namespace geode::prelude;
-
 namespace branding {
     // Optional variant of `BrandingManager`
     class BrandingManagerV2 {
     public:
-        static Result<> registerBrand(std::string const& modId, std::string const& image, BrandImageType type = BrandImageType::Sprite)
+        static geode::Result<> registerBrand(std::string modId, std::string image, BrandImageType type = BrandImageType::Sprite)
             GEODE_EVENT_EXPORT(&BrandingManagerV2::registerBrand, (modId, image, type));
     };
 };